+++ guillotina/db/strategies/__init__.py
@@ -1,4 +1,4 @@
+from . import dbresolve  # noqa
 from . import none  # noqa
-from . import novote  # noqa
 from . import resolve  # noqa
 from . import simple  # noqa